Admin client layout changes

+++ admin_client/src/views/PerformancesView.vue
@@ -12,9 +12,9 @@ const addPerformanceModal = ref<typeof AddPerformanceModal | null>()
 const goToPerformance = (id: string) => router.push(`/performances/${id}`)
 
 const handleDeletePerformance = (id: string) => {
+  if (!confirm('Are you sure you want to delete this performance?')) return
   deletePerformance(id)
   performances.value = performances.value.filter((performance) => performance._id !== id)
-  console.log(id)
 }
 
 const handlePerformanceCreated = async () => {
@@ -24,27 +24,32 @@ const handlePerformanceCreated = async () => {
 
 onMounted(async () => {
   performances.value = await getPerformances()
+  for (let i = 0; i < 4; i++) {
+    performances.value = performances.value.concat(performances.value)
+  }
 })
 </script>
 
 <template>
-  <main class="relative mt-3 space-y-1 px-2">
+  <main class="flex h-full flex-col space-y-1">
     <h1 class="text-center text-3xl font-bold">Performances</h1>
-    <button
-      @click="goToPerformance(performance._id)"
-      v-for="performance of performances"
-      :key="performance._id"
-      class="flex w-full justify-between rounded-sm border border-light p-4">
-      <div class="text-start">
-        <p class="font-bold">{{ performance.name }}</p>
-        <p>Created by: {{ performance.username }}</p>
-      </div>
-      <button @click.stop="handleDeletePerformance(performance._id)">
-        <font-awesome-icon
-          icon="fa-trash"
-          class="text-danger" />
+    <div class="flex-1 space-y-2 overflow-y-scroll">
+      <button
+        @click="goToPerformance(performance._id)"
+        v-for="performance of performances"
+        :key="performance._id"
+        class="flex w-full justify-between rounded-sm border border-light p-4">
+        <div class="text-start">
+          <p class="font-bold">{{ performance.name }}</p>
+          <p>Created by: {{ performance.username }}</p>
+        </div>
+        <button @click.stop="handleDeletePerformance(performance._id)">
+          <font-awesome-icon
+            icon="fa-trash"
+            class="text-danger" />
+        </button>
       </button>
-    </button>
+    </div>
 
     <!-- Open add performance modal -->
     <button
